+++ frontend/app/routes/_public+/_route.tsx
@@ -1,0 +1,9 @@
+import { Outlet } from '@remix-run/react';
+
+/**
+ * Do-nothing parent route.
+ * (placeholder for future code)
+ */
+export default function Route() {
+  return <Outlet />;
+}
